skal vise feilmelding for over 1000 karakterer og lage egen fil for fritekstfelt konstanter
