--- app/konstanter/sendEndringsmelding.ts
@@ -1,4 +1,0 @@
-export const SPESIAL_TEGN_REGEX = /[!@#$%^&*()?"{}|<>+¨=]/;
-export const MAKS_INPUT_LENGDE = 1000;
-export const RESPONSE_STATUS_OK = 'OK MOCK';
-export const RESPONSE_STATUS_FEIL = 'FEIL';
