+++ app/typer/fritekstfeil.ts
@@ -1,5 +1,6 @@
 export enum EFritekstFeil {
   MANGLER_TEKST = 'MANGLER_TEKST',
+  OVER_MAKS_LENGDE = 'OVER_MAKS_LENGDE',
   HAR_SPESIAL_TEGN = 'HAR_SPESIAL_TEGN',
   MINDRE_ENN_TI_TEGN = 'MINDRE_ENN_TI_TEGN',
 }
