+++ app/typer/sanity/sanitySendEndring.ts
@@ -8,6 +8,7 @@ export enum EApiKeysSendEndring {
   FRITEKSTFELT_FEILMELDING_SPESIALTEGN = 'fritekstfeltFeilmeldingSpesialTegn',
   FRITEKSTFELT_FEILMELDING_MINTEGN = 'fritekstfeltFeilmeldingMinTegn',
   FRITEKSTFELT_FEILMELDING_MANGLERTEKST = 'fritekstfeltFeilmeldingManglerTekst',
+  FRITEKSTFELT_FEILMELDING_OVERMAKSLENGDE = 'fritekstfeltFeilmeldingOverMaksLengde',
 }
 
 export type SendEndringTekstinnhold = Record<
